+++ stories/CreditAssets.stories.ts
@@ -1,0 +1,11 @@
+import CreditAssets from './CreditAssets.svelte';
+
+export default {
+  title: 'Design System/Credit Assets',
+  component: CreditAssets,
+  parameters: {
+    layout: 'fullscreen'
+  }
+};
+
+export const CreditAndPackMarks = {};